Move C API functionality to new API folder
