+++ src/api/include/GSMP_VariableManager_API.h
