+++ src/api/src/GSMP_VariableManager_API.cpp
